backend/requirements.txt
# Web Framework
fastapi==0.104.1
uvicorn==0.24.0
pydantic==2.5.0
python-multipart==0.0.6

# Database & ORM
psycopg2-binary==2.9.9
sqlalchemy==2.0.23
alembic==1.13.0

# Web Scraping
selenium==4.15.2
webdriver-manager==4.0.1
yt-dlp==2024.8.6

# Utilities
python-dotenv==1.0.0
requests==2.31.0
